Update README.md
Learn how to build a bot with Python, a library that automates your Instagram activities to get more followers and likes with minimal manual input. In this episode, he met Selenium. Your bot will open a browser, type "https://instagram.com" into the address bar, enter this with its username, and give it whatever it wants to do.